Delete engine/Optimizer.ts
